using size_t in vector

+++ engine/include/cedisplay.h
@@ -61,7 +61,7 @@ typedef struct ce_displaymng ce_displaymng;
 typedef struct {
 	void (*ctor)(ce_displaymng* displaymng, va_list args);
 	void (*dtor)(ce_displaymng* displaymng);
-	void (*enter)(ce_displaymng* displaymng, int index,
+	void (*enter)(ce_displaymng* displaymng, size_t index,
 		ce_display_rotation rotation, ce_display_reflection reflection);
 	void (*exit)(ce_displaymng* displaymng);
 } ce_displaymng_vtable;
@@ -82,7 +82,7 @@ extern void ce_displaymng_dump_supported_modes_to_stdout(ce_displaymng* displaym
 extern void ce_displaymng_dump_supported_rotations_to_stdout(ce_displaymng* displaymng);
 extern void ce_displaymng_dump_supported_reflections_to_stdout(ce_displaymng* displaymng);
 
-extern int ce_displaymng_enter(ce_displaymng* displaymng,
+extern size_t ce_displaymng_enter(ce_displaymng* displaymng,
 								int width, int height, int bpp, int rate,
 								ce_display_rotation rotation,
 								ce_display_reflection reflection);
